requirements.txt
aniso8601==9.0.1
api==0.0.7
appdirs==1.4.4
blinker==1.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
distlib==0.3.1
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
nose==1.3.7
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.19
urllib3==1.26.6
virtualenv==20.4.6
Werkzeug==2.0.1
